Restructure login & logout flows
