+++ src/pages/Home.js
@@ -1,8 +1,6 @@
 /* eslint-disable react/button-has-type */
 /* eslint-disable react/prop-types */
-
-// import SignUpCard from '../SignUpCard';
-import LoginCard from '../LoginCard';
+import SignInCard from '../SignInCard';
 
 function Container({ children }) {
   return (
@@ -23,19 +21,28 @@ function Container({ children }) {
   );
 }
 
-export default function Home() {
+export default function Home({ color, setColor, userInfo }) {
+  const boxStyle = {
+    width: '500px', border: '1px solid white', borderRadius: '5px', padding: '20px', color: 'white', textAlign: 'center', margin: '10px',
+  };
+
   return (
     <Container>
-      <div style={{
-        width: '400px', border: '1px solid white', borderRadius: '5px', padding: '20px', color: 'white', textAlign: 'center',
-      }}
-      >
+      <div style={boxStyle}>
         <h1>Feed My Mind</h1>
         <h2>A Cool Note-Taking Tool</h2>
         <p>Welcome to Feedy Notes, hope you will enjoy it!</p>
         <div>feature</div>
+        <div style={{ background: color }}>red</div>
+        <button onClick={() => setColor('blue')}>設為藍色</button>
       </div>
-      <LoginCard />
+      {userInfo ? (
+        <div style={{ color: 'white' }}>{`哈囉，${userInfo.email} ！`}</div>
+      ) : (
+        <div style={boxStyle}>
+          <SignInCard />
+        </div>
+      )}
     </Container>
   );
 }
